migrated the landing page files to the proper channel and removed it from App.jsx
